Eventloop supporting pipelining integrated to all of the code.
